requirements.txt
customtkinter==5.2.2
docx2python==3.5.0
python-docx==1.1.2
chardet
streamlit
language_tool_python
pydantic-ai
pydantic
python-dotenv
